centralize testing/fuzzing helpers
