+++ Cargo.toml
@@ -5,13 +5,17 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+default = []
+__internal-fuzz = ["libz-ng-sys"]
+
 [dependencies]
 libc = "0.2.151"
-
+libz-ng-sys = { version = "1.1.12", optional = true }
 
 [dev-dependencies]
-libz-ng-sys = "1.1.12"
 libloading = "0.8.1"
+libz-ng-sys = "1.1.12"
 
 [profile.release]
 debug = true
