+++ fuzz/Cargo.toml
@@ -19,6 +19,7 @@ libloading = "0.8.1"
 
 [dependencies.zlib]
 path = ".."
+features =  ["__internal-fuzz"]
 
 # Prevent this from interfering with workspaces
 [workspace]
